django_api/pyproject.toml
[tool.poetry]
name = "django-api"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "django_api"}]

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.5"
dj-database-url = "^1.2.0"
django = "^4.1.7"
opentelemetry-sdk = "^1.16.0"
opentelemetry-instrumentation-django = "^0.37b0"
opentelemetry-exporter-jaeger = "^1.16.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
